fix: change distance_percentage type from int to float for precision in lane calculations

+++ include/Detection/TrajectoryDefinition.hpp
@@ -53,7 +53,7 @@ class TrajectoryDefinition
     std::deque<float> recentWidths;
     const int MAX_WIDTH_HISTORY = 15;
 
-    const int distance_percentage = 0.75;
+    const float distance_percentage = 0.75;
 
     int leftLaneLastUpdatedFrame = 0;
     int rightLaneLastUpdatedFrame = 0;
